+++ PI/src/main/java/tn/esprit/pi/services/ITerrainServices.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ITerrainServices {
     Terrain addTerrain(Terrain terrain);
     Terrain updateTerrain(Terrain terrain);
-    void delete(int numterrain);
-    Terrain getById(int numterrain);
+    void delete(Long numterrain);
+    Terrain getById(Long numterrain);
     List<Terrain> getAll();
 }
